+++ R/scvelo.R
@@ -98,6 +98,31 @@
 #' }
 #' See the \pkg{scVelo} documentation for more details about the available arguments and the examples below for a syntax example.
 #'
+#' @section Supported operating systems and architectures:
+#' \pkg{scVelo} dependencies are pinned in a Conda environment to ensure reproducibility.
+#' 
+#' Differences in packages and versions available from Conda require different environments for different operating systems and architectures.
+#' \pkg{basilisk.utils} is used to determine the operating system and architecture of the computer used to run \code{scvelo()}, using to the appropriate Conda environment.
+#' 
+#' As of the latest \pkg{velociraptor} update (24 May 2024):
+#' 
+#' \describe{
+#' \item{Linux}{\pkg{scVelo} v0.3.2 from conda-forge is used.
+#' This is the latest version available to date.
+#' Note that
+#' \pkg{matplotlib} is pinned to v3.7.3 (\url{https://stackoverflow.com/questions/77128061/ydata-profiling-profilereport-attributeerror-module-matplotlib-cbook-has-no})
+#' and \pkg{numpy} is pinned to v1.23.1 (\url{https://github.com/OpenTalker/video-retalking/issues/35})}
+#' \item{MacOS Arm}{\pkg{scVelo} v0.3.2 from conda-forge is used.
+#' This is the latest version available to date.
+#' Tested on M1.}
+#' \item{Windows}{\pkg{scVelo} v0.2.5 from bioconda is used.
+#' Later versions of \pkg{scVelo} depend on \pkg{jaxlib} which is not supported on Windows (\url{https://github.com/google/jax/issues/438}).
+#' Note that \pkg{matplotlib} is pinned to v3.6.3 (\url{https://github.com/scverse/scanpy/issues/2411}),
+#' \pkg{pandas} is pinned to v1.5.2 (\url{https://stackoverflow.com/questions/76234312/importerror-cannot-import-name-is-categorical-from-pandas-api-types}),
+#' and \pkg{numpy} is pinned to v1.21.1 (\url{https://github.com/theislab/scvelo/issues/1109}).
+#' }
+#' }
+#'
 #' @return
 #' A \linkS4class{SingleCellExperiment} is returned containing the output of the velocity calculations.
 #' Of particular interest are:
@@ -124,12 +149,12 @@
 #'
 #' # make scvelo use 10 rather than the default 30 neighbors to compute moments for velocity estimation:
 #' out <- scvelo(list(X=spliced, spliced=spliced, unspliced=unspliced),
-#'               scvelo.params=list(moments=list(n_neighbors=10L)))
+#'               scvelo.params=list(neighbors=list(n_neighbors=10L)))
 #'
 #' @references
 #' Bergen, V., Lange, M., Peidli, S. et al. Generalizing RNA velocity to transient cell states through dynamical modeling. Nat Biotechnol 38, 1408–1414 (2020). \url{https://doi.org/10.1038/s41587-020-0591-3}
 #'
-#' @author Aaron Lun, Charlotte Soneson
+#' @author Aaron Lun, Charlotte Soneson, Kevin Rue-Albrecht
 #' @name scvelo
 NULL
 
@@ -138,6 +163,7 @@ NULL
 #' @importFrom BiocSingular bsparam runPCA
 #' @importFrom BiocParallel SerialParam
 #' @importFrom Matrix t
+#' @importFrom zellkonverter AnnData2SCE
 .scvelo <- function(x, subset.row=NULL,
     sf.X=NULL, sf.spliced=NULL, sf.unspliced=NULL,
     use.theirs=FALSE,
@@ -187,6 +213,7 @@ NULL
 
     and <- reticulate::import("anndata")
     scv <- reticulate::import("scvelo")
+    sc <- reticulate::import("scanpy")
     adata <- and$AnnData(X, layers=list(spliced=spliced, unspliced=unspliced))
     adata$obs_names <- rownames(spliced)
     adata$var_names <- colnames(spliced)
@@ -201,6 +228,24 @@ NULL
         do.call(scv$pp$filter_and_normalize, c(list(data=adata), scvelo.params$filter_and_normalize))
     }
 
+    do.call(sc$pp$neighbors, c(list(adata), scvelo.params$neighbors))
+    
+    if (!is.null(scvelo.params$moments)) {
+        if (!is.null(scvelo.params$moments$n_neighbors)) {
+            stop("scvelo.params$moments$n_neighbors is deprecated; use scvelo.params$neighbors$n_neighbors instead")
+        }
+        if (!is.null(scvelo.params$moments$n_pcs)) {
+            stop("scvelo.params$moments$n_pcs is deprecated; use scvelo.params$neighbors$n_pcs instead")
+        }
+    } else {
+        # if unspecified, set to NULL (= None)
+        # see https://github.com/theislab/scvelo/issues/1212
+        scvelo.params$moments <- list(
+            n_neighbors = NULL,
+            n_pcs = NULL
+        )
+    }
+    
     do.call(scv$pp$moments, c(list(data=adata), scvelo.params$moments))
 
     if (mode=="dynamical") {